voice-assistant-backend/requirements.txt
fastapi
uvicorn
faster-whisper
google-generativeai
python-dotenv
pydantic
whisper
